♻️ refactor: seperate option panel from pdf viewer for modularity
