+++ components/DataViewer/PdfViewer/index.tsx
@@ -4,8 +4,10 @@ import styles from './PdfViewer.module.css';
 import 'react-pdf/dist/Page/TextLayer.css';
 import 'react-pdf/dist/Page/AnnotationLayer.css';
 import { Dispatch, SetStateAction, useCallback, useEffect, useState } from 'react';
-import OptionPanel from './OptionPanel';
+import OptionPanel from '../OptionPanel';
 import { useViewer } from '@/states/viewer';
+import PdfViewerOptions from './PdfViewerOptions';
+import { highlightKeywordStr } from '@/utils/viewerHelper';
 
 pdfjs.GlobalWorkerOptions.workerSrc = new URL(
   'pdfjs-dist/build/pdf.worker.min.js',
@@ -13,18 +15,18 @@ pdfjs.GlobalWorkerOptions.workerSrc = new URL(
 ).toString();
 
 interface PdfViewerProps {
-  url: string;
+  dataSource: WBSourceData;
 }
 
-export default function PdfViewer({ url }: PdfViewerProps) {
-  const [tool, setTool] = useState<Tool>('SELECT');
+export default function PdfViewer({ dataSource }: PdfViewerProps) {
   const [scale, setScale] = useState<number>(1);
   const [page, setPage] = useState<number>(0);
   const [numPages, setNumPages] = useState<number>(0);
-  const { keywords, addKeyword, setDataStr } = useViewer((state) => ({
+  const { keywords, addKeyword, setDataStr, tool } = useViewer((state) => ({
     keywords: state.keywords,
     addKeyword: state.addKeyword,
     setDataStr: state.setDataStr,
+    tool: state.currentTool,
   }));
 
   useFocusKeyword(setPage);
@@ -40,19 +42,7 @@ export default function PdfViewer({ url }: PdfViewerProps) {
   const textRenderer = useCallback(
     (textItem: any) => {
       // highlights keywords
-      const pos: Pos[] = [];
-      for (const [s, f] of keywords.entries()) {
-        // find all occurences
-        if (!f) continue;
-        const matches = (textItem.str as string).matchAll(new RegExp(s, 'gi'));
-        for (const m of matches) {
-          if (m.index === undefined) continue;
-          pos.push({ start: m.index, end: m.index + s.length });
-        }
-      }
-      // add marks
-      const str = mergeAndApplyMarks(textItem.str as string, pos);
-      return str;
+      return highlightKeywordStr(textItem.str, keywords);
     },
     [keywords],
   );
@@ -62,7 +52,7 @@ export default function PdfViewer({ url }: PdfViewerProps) {
       <div className={styles.viewerContainer}>
         <div className={styles.viewer}>
           <Document
-            file={url}
+            file={dataSource.url}
             onLoadSuccess={async (d) => {
               // save all text
               const data: string[][] = [];
@@ -73,7 +63,6 @@ export default function PdfViewer({ url }: PdfViewerProps) {
                   }),
                 );
               }
-              console.log(data);
               setDataStr(data);
               setNumPages(d.numPages);
             }}
@@ -88,15 +77,15 @@ export default function PdfViewer({ url }: PdfViewerProps) {
         </div>
       </div>
 
-      <OptionPanel
-        currentTool={tool}
-        setCurrentTool={setTool}
-        scale={scale}
-        setScale={setScale}
-        page={page}
-        setPage={setPage}
-        numPages={numPages}
-      />
+      <OptionPanel>
+        <PdfViewerOptions
+          scale={scale}
+          setScale={setScale}
+          page={page}
+          setPage={setPage}
+          numPages={numPages}
+        />
+      </OptionPanel>
     </div>
   );
 }
@@ -115,62 +104,4 @@ function useFocusKeyword(setPage: Dispatch<SetStateAction<number>>) {
     });
     return () => setCallback(undefined);
   }, [setAllKeyword, setCallback, setKeyWord, setPage]);
-}
-
-interface Pos {
-  start: number;
-  end: number;
-}
-
-function posSorter(a: Pos, b: Pos) {
-  // latter item comes first
-  if (a.end > b.end) {
-    return -1;
-  } else if (a.end < b.end) {
-    return 1;
-  } else {
-    return a.start > b.start ? -1 : 1;
-  }
-}
-
-export function mergeAndApplyMarks(sourceStr: string, pos: Pos[]): string {
-  let str = sourceStr;
-  // sort occurences
-  pos.sort(posSorter);
-
-  // apply occurences
-  const bounds: Pos = { start: -1, end: -1 };
-  for (const p of pos) {
-    // first
-    if (bounds.start === -1 || bounds.end === -1) {
-      // update bounds
-      bounds.end = p.end;
-      bounds.start = p.start;
-    } else if (p.end < bounds.start) {
-      // mark
-      str =
-        str.slice(0, bounds.start) +
-        '<mark>' +
-        str.slice(bounds.start, bounds.end) +
-        '</mark>' +
-        str.slice(bounds.end);
-      // update bounds
-      bounds.end = p.end;
-      bounds.start = p.start;
-    } else {
-      bounds.end = Math.max(bounds.end, p.end);
-      bounds.start = Math.min(bounds.start, p.start);
-    }
-  }
-  // apply last mark
-  if (bounds.start !== -1 || bounds.end !== -1) {
-    str =
-      str.slice(0, bounds.start) +
-      '<mark>' +
-      str.slice(bounds.start, bounds.end) +
-      '</mark>' +
-      str.slice(bounds.end);
-  }
-
-  return str;
 }
